added data.sql implemented getSaleRecords

+++ src/main/java/com/kristovski/salesrecords/services/SaleRecordService.java
@@ -7,13 +7,14 @@ import java.util.List;
 
 public interface SaleRecordService {
 
+
     public List<SaleRecord> findAll();
 
-    public SaleRecord findById(int theId);
+    public SaleRecord findById(long theId);
 
     public void save(SaleRecord theSaleRecord);
 
-    public void deleteById(int theId);
+    public void deleteById(long theId);
 
     public List<SaleRecord> searchBy(LocalDate theDate);
